+++ Source/EventFlow/Aggregates/IAggregateRoot.cs
@@ -28,13 +28,12 @@ using EventFlow.EventStores;
 
 namespace EventFlow.Aggregates
 {
-    public interface IAggregateRoot<out TIdentity>
-        where TIdentity : IIdentity
+    public interface IAggregateRoot
     {
-        TIdentity Id { get; }
+        IAggregateName Name { get; }
         int Version { get; }
-        bool IsNew { get; }
         IEnumerable<IAggregateEvent> UncommittedEvents { get; }
+        bool IsNew { get; }
 
         Task<IReadOnlyCollection<IDomainEvent>> CommitAsync(
             IEventStore eventStore,
@@ -44,5 +43,11 @@ namespace EventFlow.Aggregates
         void ApplyEvents(IEnumerable<IAggregateEvent> aggregateEvents);
 
         void ApplyEvents(IReadOnlyCollection<IDomainEvent> domainEvents);
+    }
+
+    public interface IAggregateRoot<out TIdentity> : IAggregateRoot
+        where TIdentity : IIdentity
+    {
+        TIdentity Id { get; }
     }
 }