+++ Source/EventFlow/Configuration/Registrations/AutofacDecorator.cs
@@ -34,20 +34,19 @@ namespace EventFlow.Configuration.Registrations
         {
             if (level < 0)
             {
-                throw new ArgumentOutOfRangeException("level");
+                throw new ArgumentOutOfRangeException(nameof(level));
             }
 
             return level == 0
                 ? ServiceType.FullName
-                : string.Format("{0} (level {1})", ServiceType.FullName, level);
+                : $"{ServiceType.FullName} (level {level})";
         }
     }
 
     internal class AutofacDecorator<TService> : AutofacDecorator
     {
         private readonly Func<IResolverContext, TService, TService> _factory;
-        private readonly Type _serviceType = typeof(TService);
-        public override Type ServiceType { get { return _serviceType; } }
+        public override Type ServiceType { get; } = typeof(TService);
 
         public AutofacDecorator(Func<IResolverContext, TService, TService> factory)
         {