+++ Source/EventFlow/Core/Identity.cs
@@ -23,6 +23,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 using System.Text.RegularExpressions;
 using EventFlow.ValueObjects;
 
@@ -32,25 +33,47 @@ namespace EventFlow.Core
         where T : Identity<T>
     {
         // ReSharper disable StaticMemberInGenericType
-        private static readonly Regex NameReplace = new Regex("Id$", RegexOptions.Compiled);
-        private static readonly string Name = NameReplace.Replace(typeof (T).Name, string.Empty).ToLowerInvariant();
-        private static readonly Regex ValueValidation = new Regex(
-            @"^[a-z0-9]+\-[a-f0-9]{8}\-[a-f0-9]{4}\-[a-f0-9]{4}\-[a-f0-9]{4}\-[a-f0-9]{12}$",
-            RegexOptions.Compiled);
+        private static readonly string Name;
+        private static readonly Regex ValueValidation;
         // ReSharper enable StaticMemberInGenericType
 
-        public static T New
+        static Identity()
         {
-            get
-            {
-                var value = string.Format("{0}-{1}", Name, Guid.NewGuid()).ToLowerInvariant();
-                return With(value);
-            }
+            var nameReplace = new Regex("Id$");
+            Name = nameReplace.Replace(typeof (T).Name, string.Empty).ToLowerInvariant();
+            ValueValidation = new Regex(
+                @"^[a-z0-9]+\-[a-f0-9]{8}\-[a-f0-9]{4}\-[a-f0-9]{4}\-[a-f0-9]{4}\-[a-f0-9]{12}$",
+                RegexOptions.Compiled);
+        }
+
+        public static T New => BuildWith(Guid.NewGuid());
+
+        public static T NewDeterministic(Guid namespaceId, string name)
+        {
+            var guid = GuidFactories.Deterministic.Create(namespaceId, name);
+            return BuildWith(guid);
         }
 
         public static T With(string value)
         {
-            return (T)Activator.CreateInstance(typeof(T), value);
+            try
+            {
+                return (T)Activator.CreateInstance(typeof(T), value);
+            }
+            catch (TargetInvocationException e)
+            {
+                if (e.InnerException != null)
+                {
+                    throw e.InnerException;
+                }
+                throw;
+            }
+        }
+
+        private static T BuildWith(Guid guid)
+        {
+            var value = $"{Name}-{guid}".ToLowerInvariant();
+            return With(value);
         }
 
         public static bool IsValid(string value)
@@ -62,16 +85,16 @@ namespace EventFlow.Core
         {
             if (string.IsNullOrEmpty(value))
             {
-                yield return string.Format("Aggregate ID of type '{0}' is null or empty", typeof(T).Name);
+                yield return $"Aggregate ID of type '{typeof (T).Name}' is null or empty";
                 yield break;
             }
 
             if (!string.Equals(value.Trim(), value, StringComparison.InvariantCulture))
-                yield return string.Format("Aggregate ID '{0}' of type '{1}' contains leading and/or traling spaces", value, typeof(T).Name);
+                yield return $"Aggregate ID '{value}' of type '{typeof (T).Name}' contains leading and/or traling spaces";
             if (!value.StartsWith(Name))
-                yield return string.Format("Aggregate ID '{0}' of type '{1}' does not start with '{2}'", value, typeof(T).Name, Name);
+                yield return $"Aggregate ID '{value}' of type '{typeof (T).Name}' does not start with '{Name}'";
             if (!ValueValidation.IsMatch(value))
-                yield return string.Format("Aggregate ID '{0}' of type '{1}' does not follow the syntax '[NAME]-[GUID]' in lower case", value, typeof(T).Name);
+                yield return $"Aggregate ID '{value}' of type '{typeof (T).Name}' does not follow the syntax '[NAME]-[GUID]' in lower case";
         }
 
         protected Identity(string value) : base(value)